+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added `TSMiniWebBrowser:showToolBar`, `opaque` and `backgroundColor` - [@dblock](https://github.com/dblock).
 * Renamed `TSMiniWebBrowser:initWithUrl` to `initWithURL` - [@dblock](https://github.com/dblock).
 * Exposed `TSMiniWebBrowser:requestWithURL:(NSURL *)url` for subclasses of `TSMiniWebBrowser` - [@dblock](https://github.com/dblock).
+* Exposed `(NSURL *)TSMiniWebBrowser:currentURL` which is no longer an NSString and is also set after `initWithURL` - [@dblock](https://github.com/dblock).
 
 ### 1.1 (5/12/2014)